Update README.md
Abstract. The next generation of tightly interconnected vehicles offers
a variety of new technologies as well as business opportunities where vehicles
form so-called vehicular ad hoc networks (VANETs). Sophisticated
systems of interconnected cars in the context of VANETs often assume
a full coverage of 5G networks to unfold their full potential. However,
nowadays network infrastructure is neither 5G-ready nor does it provide
reliable and full coverage of all areas that might be relevant for VANETbased
networks. Hence, the conventional mode of operations of all nodes
being connected to one blockchain at all times is not feasible. In addition,
traditional blockchains require every node to process each transaction
and smart contract command which is highly inefficient. Therefore,
we present a solution based on mobile ad hoc blockchains that enable
groups of nodes involved in any kind of collaboration to effectively form
temporary networks and coordinate themselves.
Another critical requirement of VANETs and interconnected vehicles is
security. The safety of network participants does not only depend on
the vehicle’s hardware, but also on the correctness of the software that
controls the interaction and transaction within the network. Formal verification
is a common way to address the issue proving the correctness of
software. The self-amending blockchain Tezos supports Turing-complete
smart contracts and also offers built-in formal verification of their programming
languages, thereby fostering the security of our solution.
This whitepaper fills the gap by introducing Tezos-based vehicular ad
hoc blockchains based on mobile ad hoc blockchains that allow groups of
nodes to be temporarily disconnected from the overall network but still
being able to enact and transact on a local network level for the duration
of their interaction. We present the advantages of the system, outline the
requirements and goals, as well as the architecture of the system.